refactor(compiler): ♻️ extract reusable Arena into compiler/support/
- Move duplicated bump allocator from AstContext and TypeContext into
  a shared Arena class at compiler/support/arena.h
- AstContext and TypeContext now compose Arena instead of duplicating it
- Add compiler/support/ to ARCH_INDEX.md
