+++ compiler/frontend/ast/ast.h
@@ -2,12 +2,10 @@
 #define DAO_FRONTEND_AST_AST_H
 
 #include "frontend/diagnostics/source.h"
+#include "support/arena.h"
 
 #include <cassert>
 #include <cstdint>
-#include <functional>
-#include <memory>
-#include <ranges>
 #include <string_view>
 #include <utility>
 #include <vector>
@@ -15,89 +13,25 @@
 namespace dao {
 
 // ---------------------------------------------------------------------------
-// Arena allocator — owns all AST nodes for a compilation unit.
+// AstContext — owns all AST nodes for a compilation unit.
 // ---------------------------------------------------------------------------
 
 class AstContext {
 public:
   AstContext() = default;
-  ~AstContext() {
-    // Destruct in reverse allocation order.
-    for (auto& dtor : std::ranges::reverse_view(dtors_)) {
-      dtor();
-    }
-    for (auto* block : blocks_) {
-      ::operator delete(block);
-    }
-  }
+  ~AstContext() = default;
 
   AstContext(const AstContext&) = delete;
   auto operator=(const AstContext&) -> AstContext& = delete;
-
-  AstContext(AstContext&& other) noexcept
-      : blocks_(std::move(other.blocks_)), offset_(other.offset_), dtors_(std::move(other.dtors_)) {
-    other.blocks_.clear();
-    other.offset_ = kBlockSize;
-    other.dtors_.clear();
-  }
-
-  auto operator=(AstContext&& other) noexcept -> AstContext& {
-    if (this != &other) {
-      // Destroy current contents first.
-      for (auto& dtor : std::ranges::reverse_view(dtors_)) {
-        dtor();
-      }
-      for (auto* block : blocks_) {
-        ::operator delete(block);
-      }
-      blocks_ = std::move(other.blocks_);
-      offset_ = other.offset_;
-      dtors_ = std::move(other.dtors_);
-      other.blocks_.clear();
-      other.offset_ = kBlockSize;
-      other.dtors_.clear();
-    }
-    return *this;
-  }
+  AstContext(AstContext&&) noexcept = default;
+  auto operator=(AstContext&&) noexcept -> AstContext& = default;
 
   template <typename T, typename... Args> auto alloc(Args&&... args) -> T* {
-    void* mem = allocate(sizeof(T), alignof(T));
-    auto* ptr = new (mem) T(std::forward<Args>(args)...);
-    if constexpr (!std::is_trivially_destructible_v<T>) {
-      dtors_.push_back([ptr]() { ptr->~T(); }); // NOLINT(modernize-use-trailing-return-type)
-    }
-    return ptr;
+    return arena_.alloc<T>(std::forward<Args>(args)...);
   }
 
 private:
-  static constexpr size_t kBlockSize = 4096;
-
-  struct Block {
-    char data[kBlockSize]; // NOLINT(modernize-avoid-c-arrays)
-  };
-
-  std::vector<Block*> blocks_;
-  size_t offset_ = kBlockSize; // Force first allocation to create a block.
-  std::vector<std::function<void()>> dtors_;
-
-  auto allocate(size_t size, size_t align) -> void* {
-    // Align up.
-    offset_ = (offset_ + align - 1) & ~(align - 1);
-    if (offset_ + size > kBlockSize) {
-      if (size > kBlockSize) {
-        // Oversized allocation — give it its own block.
-        auto* mem = ::operator new(size);
-        blocks_.push_back(static_cast<Block*>(mem));
-        return mem;
-      }
-      blocks_.push_back(new Block);
-      offset_ = 0;
-    }
-    void* ptr =
-        blocks_.back()->data + offset_; // NOLINT(cppcoreguidelines-pro-bounds-pointer-arithmetic)
-    offset_ += size;
-    return ptr;
-  }
+  Arena arena_;
 };
 
 // ---------------------------------------------------------------------------
