+++ compiler/frontend/types/type_context.cpp
@@ -2,42 +2,8 @@
 
 #include <cstdint>
 #include <functional>
-#include <ranges>
 
 namespace dao {
-
-// ---------------------------------------------------------------------------
-// Arena helpers
-// ---------------------------------------------------------------------------
-
-auto TypeContext::allocate(size_t size, size_t align) -> void* {
-  offset_ = (offset_ + align - 1) & ~(align - 1);
-  if (offset_ + size > kBlockSize) {
-    if (size > kBlockSize) {
-      auto* mem = ::operator new(size);
-      blocks_.push_back(static_cast<Block*>(mem));
-      return mem;
-    }
-    blocks_.push_back(new Block);
-    offset_ = 0;
-  }
-  void* ptr =
-      blocks_.back()->data + offset_; // NOLINT(cppcoreguidelines-pro-bounds-pointer-arithmetic)
-  offset_ += size;
-  return ptr;
-}
-
-void TypeContext::destroy() {
-  for (auto& dtor : std::ranges::reverse_view(dtors_)) {
-    dtor();
-  }
-  for (auto* block : blocks_) {
-    ::operator delete(block);
-  }
-  blocks_.clear();
-  offset_ = kBlockSize;
-  dtors_.clear();
-}
 
 // ---------------------------------------------------------------------------
 // Hash helpers
@@ -83,46 +49,19 @@ auto TypeContext::GenericParamKeyHash::operator()(const GenericParamKey& key) co
 // ---------------------------------------------------------------------------
 
 TypeContext::TypeContext() {
-  // Pre-populate all builtin types.
+  // Pre-populate all builtin scalar types.
   for (uint8_t i = 0; i < kBuiltinKindCount; ++i) {
     auto kind = static_cast<BuiltinKind>(i);
-    builtins_[i] = alloc<TypeBuiltin>(kind);
+    builtins_[i] = arena_.alloc<TypeBuiltin>(kind);
   }
+  // Void is a compiler-internal singleton, not a builtin scalar.
+  void_ = arena_.alloc<TypeVoid>();
 }
 
-TypeContext::~TypeContext() { destroy(); }
+TypeContext::~TypeContext() = default;
 
-TypeContext::TypeContext(TypeContext&& other) noexcept
-    : blocks_(std::move(other.blocks_)), offset_(other.offset_),
-      dtors_(std::move(other.dtors_)), builtins_(other.builtins_),
-      pointer_map_(std::move(other.pointer_map_)),
-      function_map_(std::move(other.function_map_)),
-      named_map_(std::move(other.named_map_)),
-      generic_param_map_(std::move(other.generic_param_map_)) {
-  other.blocks_.clear();
-  other.offset_ = kBlockSize;
-  other.dtors_.clear();
-  other.builtins_ = {};
-}
-
-auto TypeContext::operator=(TypeContext&& other) noexcept -> TypeContext& {
-  if (this != &other) {
-    destroy();
-    blocks_ = std::move(other.blocks_);
-    offset_ = other.offset_;
-    dtors_ = std::move(other.dtors_);
-    builtins_ = other.builtins_;
-    pointer_map_ = std::move(other.pointer_map_);
-    function_map_ = std::move(other.function_map_);
-    named_map_ = std::move(other.named_map_);
-    generic_param_map_ = std::move(other.generic_param_map_);
-    other.blocks_.clear();
-    other.offset_ = kBlockSize;
-    other.dtors_.clear();
-    other.builtins_ = {};
-  }
-  return *this;
-}
+TypeContext::TypeContext(TypeContext&&) noexcept = default;
+auto TypeContext::operator=(TypeContext&&) noexcept -> TypeContext& = default;
 
 // ---------------------------------------------------------------------------
 // Builtin access
@@ -132,6 +71,10 @@ auto TypeContext::builtin(BuiltinKind kind) -> const TypeBuiltin* {
   return builtins_[static_cast<uint8_t>(kind)];
 }
 
+auto TypeContext::void_type() -> const TypeVoid* {
+  return void_;
+}
+
 // ---------------------------------------------------------------------------
 // Interned constructors
 // ---------------------------------------------------------------------------
@@ -139,7 +82,7 @@ auto TypeContext::builtin(BuiltinKind kind) -> const TypeBuiltin* {
 auto TypeContext::pointer_to(const Type* pointee) -> const TypePointer* {
   auto [it, inserted] = pointer_map_.try_emplace(pointee, nullptr);
   if (inserted) {
-    it->second = alloc<TypePointer>(pointee);
+    it->second = arena_.alloc<TypePointer>(pointee);
   }
   return it->second;
 }
@@ -149,7 +92,7 @@ auto TypeContext::function_type(std::vector<const Type*> params,
   FnKey key{params, ret};
   auto [it, inserted] = function_map_.try_emplace(key, nullptr);
   if (inserted) {
-    it->second = alloc<TypeFunction>(std::move(params), ret);
+    it->second = arena_.alloc<TypeFunction>(std::move(params), ret);
   }
   return it->second;
 }
@@ -160,7 +103,7 @@ auto TypeContext::named_type(const void* decl_id, std::string_view name,
   NamedKey key{decl_id, type_args};
   auto [it, inserted] = named_map_.try_emplace(key, nullptr);
   if (inserted) {
-    it->second = alloc<TypeNamed>(decl_id, name, std::move(type_args));
+    it->second = arena_.alloc<TypeNamed>(decl_id, name, std::move(type_args));
   }
   return it->second;
 }
@@ -170,7 +113,7 @@ auto TypeContext::generic_param(std::string_view name, uint32_t index)
   GenericParamKey key{name, index};
   auto [it, inserted] = generic_param_map_.try_emplace(key, nullptr);
   if (inserted) {
-    it->second = alloc<TypeGenericParam>(name, index);
+    it->second = arena_.alloc<TypeGenericParam>(name, index);
   }
   return it->second;
 }
@@ -182,13 +125,13 @@ auto TypeContext::generic_param(std::string_view name, uint32_t index)
 auto TypeContext::make_struct(const void* decl_id, std::string_view name,
                               std::vector<StructField> fields)
     -> const TypeStruct* {
-  return alloc<TypeStruct>(decl_id, name, std::move(fields));
+  return arena_.alloc<TypeStruct>(decl_id, name, std::move(fields));
 }
 
 auto TypeContext::make_enum(const void* decl_id, std::string_view name,
                              std::vector<EnumVariant> variants)
     -> const TypeEnum* {
-  return alloc<TypeEnum>(decl_id, name, std::move(variants));
+  return arena_.alloc<TypeEnum>(decl_id, name, std::move(variants));
 }
 
 } // namespace dao
